requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
scikit-learn==1.3.0
pytest==7.4.0
pytest-asyncio==0.21.1
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.7.0